test: cover generate edges through the handler per review
