+++ src/handlers/project/add/policy/index.ts
@@ -2,7 +2,8 @@ import z from "zod";
 import { InputValidationError } from "../../../../errors";
 import { SourceResolver } from "../../../../io";
 import { gatewayResourceName } from "../../../../projectSchemas/gateway";
-import { policyEngineResourceName, type PolicySchema } from "../../../../projectSchemas/policy";
+import type { PolicySchema } from "../../../../projectSchemas/policy";
+import { policyEngineResourceName } from "../policy-engine";
 import { createHandler, flag, ProjectKey } from "../../../../router";
 import { coreOptsFromCtx } from "../../../utils";
 import type { AddProjectResourceConfig } from "../types";
